rails-ruby extensions are optional and not enabled by default
Many don't desire the modifications rails-minor-mode makes on ruby-mode.
It seems to be reaching out of it's scope.  Rails-minor-mode should
focus on providing helpers that specifically contribute to rails, and
not needlessly modifying behavior of ruby-mode.  Extensions that modify
ruby-mode should belong to another package, or as an optional include.